# backend/api/oas-file.yaml
openapi: 3.0.3
info:
  title: AppSciArt API
  version: 1.0.0
  description: API interactions required by AppSciArt platform.
security:
  - bearerAuth: []
...
paths:
...
  /hackathons:
    x-router-controller: hackathonsController
    get:
      security: [{}]
      tags: [Hackathons]
      summary: Get hackathons
      operationId: getHackathons
      parameters:
        - name: filter
          in: query
          required: true
          schema:
            type: string
            enum: [closest, incoming, previous, all]
      responses:
        "200":
          description: List of hackathons
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/HackathonList"
                  - $ref: "#/components/schemas/Hackathon"
        default:
          description: Unexpected error

    post:
      tags: [Hackathons]
      summary: Create a new hackathon
      operationId: createHackathon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Hackathon"
      responses:
        "201":
          description: Hackathon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hackathon"
  /hackathons/{hackathonId}:
    x-router-controller: hackathonsController
    get:
      tags: [Hackathons]
      summary: Get details of a hackathon
      operationId: getHackathonDetails
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Hackathon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hackathon"
        "401":
          $ref: '#/components/responses/UnauthorisedError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    put:
      tags: [Hackathons]
      summary: Update a hackathon
      operationId: updateHackathon
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Hackathon"
      responses:
        "200":
          description: Hackathon updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hackathon"
    delete:
      tags: [Hackathons]
      summary: Delete a hackathon
      operationId: deleteHackathon
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Hackathon deleted
  /hackathons/{hackathonId}/publish:
    x-router-controller: hackathonsController
    patch:
      tags: [Hackathons]
      summary: Publish a hackathon
      operationId: publishHackathon
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Hackathon published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hackathon"
  /hackathons/{hackathonId}/users:
    x-router-controller: hackathonsController
    get:
      tags: [Hackathons]
      summary: Get users of a hackathon
      operationId: getHackathonUsers
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserList"
...
  /seeds:
    x-router-controller: seedsController
    get:
      security: [{}]
      tags: [SciArt seeds]
      summary: Get seeds
      operationId: getSeeds
      parameters:
        - name: editionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: hackathonId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of seeds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedList"
        default:
          $ref: '#/components/responses/UnexpectedError'
    post:
      tags: [SciArt seeds]
      summary: Create a new seed
      operationId: createSeed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedDetails"
      responses:
        "201":
          description: Seed created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedDetails"
  /flowers:
    x-router-controller: flowersController
    get:
      security: [{}]
      tags: [SciArt flowers]
      summary: Get flowers
      operationId: getFlowersByEdition
      parameters:
        - name: editionId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of flowers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowerList"
        default:
          $ref: '#/components/responses/UnexpectedError'
    post:
      tags: [SciArt flowers]
      summary: Create a new flower
      operationId: createFlower
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowerDetails"
      responses:
        "201":
          description: Flower created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowerDetails"
  /fruits:
    x-router-controller: fruitsController
    get:
      security: [{}]
      tags: [SciArt fruits]
      summary: Get fruits
      operationId: getFruitsByEdition
      parameters:
        - name: editionId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of fruits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FruitList"
        default:
          $ref: '#/components/responses/UnexpectedError'
    post:
      tags: [SciArt fruits]
      summary: Create a new fruit
      operationId: createFruit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FruitDetails"
      responses:
        "201":
          description: Fruit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FruitDetails"
  /seeds/{seedId}:
    x-router-controller: seedsController
    get:
      security: [{}]
      tags: [SciArt seeds]
      summary: Get seed details
      operationId: getSeedDetails
      parameters:
        - name: seedId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Seed details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedDetails"
        "401":
          $ref: '#/components/responses/UnauthorisedError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    put:
      tags: [SciArt seeds]
      summary: Update a seed
      operationId: updateSeed
      parameters:
        - name: seedId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedDetails"
      responses:
        "200":
          description: Seed updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedDetails"
    delete:
      tags: [SciArt seeds]
      summary: Delete a seed
      operationId: deleteSeed
      parameters:
        - name: seedId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Seed deleted
  /flowers/{flowerId}:
    x-router-controller: flowersController
    get:
      security: [{}]
      tags: [SciArt flowers]
      summary: Get flower details
      operationId: getFlowerDetails
      parameters:
        - name: flowerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Flower details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowerDetails"
        "401":
          $ref: '#/components/responses/UnauthorisedError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    put:
      tags: [SciArt flowers]
      summary: Update a flower
      operationId: updateFlower
      parameters:
        - name: flowerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowerDetails"
      responses:
        "200":
          description: Flower updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowerDetails"
    delete:
      tags: [SciArt flowers]
      summary: Delete a flower
      operationId: deleteFlower
      parameters:
        - name: flowerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Flower deleted
  /flowers/{flowerId}/publish:
    x-router-controller: flowersController
    patch:
      tags: [SciArt flowers]
      summary: Publish a flower
      operationId: publishFlower
      parameters:
        - name: flowerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Flower published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowerDetails"
  /seeds/{seedId}/publish:
    x-router-controller: seedsController
    patch:
      tags: [SciArt seeds]
      summary: Publish a seed
      operationId: publishSeed
      parameters:
        - name: seedId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Seed published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedDetails"
  /fruits/{fruitId}:
    x-router-controller: fruitsController
    get:
      security: [{}]
      tags: [SciArt fruits]
      summary: Get fruit details
      operationId: getFruitDetails
      parameters:
        - name: fruitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Fruit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FruitDetails"
        "401":
          $ref: '#/components/responses/UnauthorisedError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    put:
      tags: [SciArt fruits]
      summary: Update a fruit
      operationId: updateFruit
      parameters:
        - name: fruitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FruitDetails"
      responses:
        "200":
          description: Fruit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FruitDetails"
    delete:
      tags: [SciArt fruits]
      summary: Delete a fruit
      operationId: deleteFruit
      parameters:
        - name: fruitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Fruit deleted
  /fruits/{fruitId}/publish:
    x-router-controller: fruitsController
    patch:
      tags: [SciArt fruits]
      summary: Publish a fruit
      operationId: publishFruit
      parameters:
        - name: fruitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Fruit published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FruitDetails"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorisedError:
      description: Unauthorised
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SeedDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        name:
          type: string
          maxLength: 255
        templateLink:
          type: string
          format: uri
          nullable: true
        imageLink:
          type: string
          format: uri
          nullable: true
        state:
          type: string
          enum: [IN_BLANK, IN_PROGRESS, IN_REVIEW, PUBLISHED]
        branchesOfKnowledge:
          type: array
          items:
            type: string
            maxLength: 50
          nullable: true
        videoLink:
          type: string
          format: uri
          nullable: true
        presentationLink:
          type: string
          format: uri
          nullable: true
        podcastLink:
          type: string
          format: uri
          nullable: true
        resourcesLink:
          type: string
          format: uri
          nullable: true
        authors:
          type: array
          items:
            $ref: "#/components/schemas/UserDetails"
          nullable: true
        editionId:
          type: string
          format: uuid
          minimum: 1
        likeCount:
          type: integer
          minimum: 0
          nullable: true
    FlowerDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        name:
          type: string
          maxLength: 255
        templateLink:
          type: string
          format: uri
        state:
          type: string
          enum: [IN_BLANK, IN_PROGRESS, IN_REVIEW, PUBLISHED]
        concept:
          type: string
          maxLength: 1500
          nullable: true
        resourcesLink:
          type: string
          format: uri
        seedId:
          type: string
          format: uuid
          minimum: 1
        authors:
          type: array
          items:
            $ref: "#/components/schemas/UserDetails"
        hackathonId:
          type: string
          format: uuid
          minimum: 1
    FruitDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        title:
          type: string
          maxLength: 255
        mainImage:
          type: string
          format: uri
          nullable: true
        state:
          type: string
          enum: [IN_BLANK, IN_PROGRESS, IN_REVIEW, PUBLISHED]
        authorVision:
          type: string
          maxLength: 1500
          nullable: true
        curatorVision:
          type: string
          maxLength: 1500
          nullable: true
        driveLink:
          type: string
          format: uri
          nullable: true
        flowerId:
          type: string
          format: uuid
          nullable: true
    UserDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 50
        surname:
          type: string
          maxLength: 50
        password:
          type: string
          format: password
        gender:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
    UserList:
      type: array
      items:
        $ref: "#/components/schemas/UserDetails"
    SeedList:
      type: array
      items:
        $ref: "#/components/schemas/SeedDetails"
    FlowerList:
      type: array
      items:
        $ref: "#/components/schemas/FlowerDetails"
    FruitList:
      type: array
      items:
        $ref: "#/components/schemas/FruitDetails"
    Hackathon:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minimum: 1
        logo:
          type: string
          format: uri
          nullable: true
        type:
          type: string
          enum: [ON_SITE, ONLINE, HYBRID]
        location:
          type: string
          maxLength: 255
          nullable: true
        description:
          type: string
          maxLength: 500
          nullable: true
        editionId:
          type: string
          format: uuid
          minimum: 1
        isVisible:
          type: boolean
        meetLink:
          type: string
          format: uri
          nullable: true
    HackathonList:
      type: array
      items:
        $ref: "#/components/schemas/Hackathon"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Internal Server Error