# backend/src/controllers/flowersController.js
import * as service from '../services/flowersService.js'
import { validateIsPublishedOrStaff } from '../validators/productValidators.js'
import { checkIsStaff } from '../validators/userValidators.js'
import { withErrorHandler } from './errorHandling.js'
import { validateEditionById } from '../validators/hackathonAndEditionValidators.js'

export const getFlowersByEdition = withErrorHandler(async (req, res) => {
  const editionId = req.query.editionId
  await validateEditionById(req, editionId)

  const showUnpublished = await checkIsStaff(req)
  const flowers = await service.getFlowersByEdition(editionId, showUnpublished)
  return res.status(200).send(flowers)
})

export function createFlower (req, res) {
  service.createFlower(req, res)
}

export const getFlowerDetails = withErrorHandler(async (req, res) => {
  const flowerId = req.params.flowerId
  const flower = await service.getFlowerDetails(flowerId)
  await validateIsPublishedOrStaff(req, flower)
  return res.status(200).send(flower)
})

export function updateFlower (req, res) {
  service.updateFlower(req, res)
}

export function deleteFlower (req, res) {
  service.deleteFlower(req, res)
}

export function publishFlower (req, res) {
  service.publishFlower(req, res)
}